+++ delineation_raster.R
@@ -2,7 +2,7 @@
 # Description: Computes urban delineations from raster
 # Author: Clement Gorin
 # Contact: gorinclem@gmail.com
-# Updated: 08 July 2021
+# Version: 2021.10.26
 
 # Packages
 suppressMessages(if(!require('pacman')) install.packages('pacman', repos = 'https://cloud.r-project.org/'))
@@ -32,8 +32,8 @@ params <- parse_args(OptionParser(usage = 'Computes urban delineations from rast
 params <- subset(params, names(params) != 'help')
 
 # (!) Testing only --------------------------------------------------------
-# setwd("~/Desktop/delineation")
-# # Parameters
+# setwd('~/Desktop/delineation')
+# # Interactive
 # params <- modifyList(params, list(
 #   density   = 'input/raster/density.tif',
 #   unlivable = 'input/raster/unlivable.tif',
@@ -47,7 +47,7 @@ params <- subset(params, names(params) != 'help')
 # Checks directories
 if(!dir.exists(params$outdir)) dir.create(params$outdir)
 if(!dir.exists(params$tmpdir)) dir.create(params$tmpdir)
-unlink(dir(params$tmpdir, pattern = "^boot_.*\\.fst$", full.names = T), recursive = T)
+unlink(dir(params$tmpdir, pattern = '^boot_.*\\.fst$', full.names = T), recursive = T)
 
 # Checks parameters
 tests <- list(
@@ -79,11 +79,11 @@ params$memory  <- ifelse(params$memory  == -1, maxmem, min(params$memory,  maxme
 rm(maxcor, maxmem)
 
 # Prints parameters
-cat('\nComputes urban delineations from raster (version 21-07-08)\n')
+cat('\nComputes urban delineations from raster (version 2021.10.26)\n')
 cat('\nParameters:', sprintf('- %-10s= %s', names(params), unlist(params)), sep = '\n')
 
 # Sets up workers
-cat('\nOperations:\n- Optimising computations')
+cat('\nOperations:')
 plan(multiprocess, workers = params$workers, gc = T)
 
 # Functions ---------------------------------------------------------------
@@ -100,7 +100,7 @@ read_foo <- cmpfun(function(file, livable = NULL) {
   image <- raster(file)
   image <- as.cimg(image, maxpixels = ncell(image))
   if(!is.null(livable)) {
-    image <- pad(image, 2 * ceiling(params$bandwidth / 2), "xy")
+    image <- pad(image, 2 * ceiling(params$bandwidth / 2), 'xy')
     image <- replace(image, !livable, 0)
   }
   return(image)
@@ -116,7 +116,7 @@ filename_foo <- cmpfun(function(label, params) {
 # Writes cimg as raster
 write_foo <- cmpfun(function(image, label, params, navalue = -1) {
   container <- raster(params$density) < 0
-  image     <- crop.borders(image, nPix = floor((params$bandwidth + 1) / 2))
+  image     <- crop.borders(image, nPix = ((nrow(image) - ncol(container)) / 2))
   image     <- setValues(container, c(image))
   image     <- mask(image, container)
   filename  <- filename_foo(label, params)
@@ -145,26 +145,51 @@ kernel_foo <- cmpfun(function(bandwidth) {
 })
 
 # Computes single bootstrap
-bootstrap_foo <- cmpfun(function(density, livable, rescale, kernel, params) {
+bootstrap_foo <- cmpfun(function(density, livable, kernel, params) {
   bootstrap <- replace(density, livable, sample(density[livable], replace = params$replace))
-  bootstrap <- replace(bootstrap, livable, bootstrap[livable] / rescale[livable])
   bootstrap <- convolve(bootstrap, kernel)
   bootstrap <- subset(bootstrap, livable)
   return(bootstrap)
 })
 
+# Computes single bootstrap for cores
+bootstrap_cores_foo <- cmpfun(function(density1, livable1, density2, livable2, kernel, params) {
+  bootstrap1 <- replace(density1, livable1, sample(density1[livable1], replace = params$replace))
+  bootstrap2 <- replace(density2, livable2, sample(density2[livable2], replace = params$replace))
+  bootstrap  <- bootstrap1 + bootstrap2
+  bootstrap  <- convolve(bootstrap, kernel)
+  bootstrap  <- subset(bootstrap, livable)
+  return(bootstrap)
+})
+
 # Computes multiple bootstraps
-bootstraps_foo <- cmpfun(function(density, livable, rescale, kernel, params) {
+bootstraps_foo <- cmpfun(function(density, livable, kernel, params) {
   if(params$usedisk) {
     bootstraps <- future_sapply(1:params$nboots, function(.) {
-      bootstrap <- bootstrap_foo(density, livable, rescale, kernel, params)
+      bootstrap <- bootstrap_foo(density, livable, kernel, params)
       file      <- tempfile('boot_', params$tmpdir, '.fst')
       write_fst(data.frame(bootstrap), file, compress = 0)
       gc()
       return(file)
     }, future.seed = params$seed)
   } else {
-    bootstraps <- future_replicate(params$nboots, bootstrap_foo(density, livable, rescale, kernel, params), future.seed = params$seed)
+    bootstraps <- future_replicate(params$nboots, bootstrap_foo(density, livable, kernel, params), future.seed = params$seed)
+  }
+  return(bootstraps)
+})
+
+# Computes multiple bootstraps for cores
+bootstraps_cores_foo <- cmpfun(function(density1, livable1, density2, livable2, kernel, params) {
+  if(params$usedisk) {
+    bootstraps <- future_sapply(1:params$nboots, function(.) {
+      bootstrap <- bootstrap_cores_foo(density1, livable1, density2, livable2, kernel, params)
+      file      <- tempfile('boot_', params$tmpdir, '.fst')
+      write_fst(data.frame(bootstrap), file, compress = 0)
+      gc()
+      return(file)
+    }, future.seed = params$seed)
+  } else {
+    bootstraps <- future_replicate(params$nboots, bootstrap_cores_foo(density1, livable1, density2, livable2, kernel, params), future.seed = params$seed)
   }
   return(bootstraps)
 })
@@ -181,15 +206,14 @@ threshold_foo <- cmpfun(function(bootstraps, livable, params) {
     threshold <- do.call(c, threshold)
   } else {
     threshold <- rowQuantiles(bootstraps, probs = (params$quantile / 100))
-  }
+  } 
   threshold <- replace(as.cimg(livable), livable, threshold)
   return(threshold)
 })
 
 # Computes delineations
-delineation_foo <- cmpfun(function(density, livable, rescale, threshold, kernel) {
-  delineation <- replace(density, livable, density[livable] / rescale[livable])
-  delineation <- convolve(delineation, kernel)
+delineation_foo <- cmpfun(function(density, livable, threshold, kernel) {
+  delineation <- convolve(density, kernel)
   delineation <- replace(delineation, !livable, 0)
   delineation <- as.cimg(delineation > threshold)
   return(delineation)
@@ -197,12 +221,11 @@ delineation_foo <- cmpfun(function(density, livable, rescale, threshold, kernel)
 
 # Computes ranks
 rank_foo <- cmpfun(function(identifier) {
-    values     <- subset(identifier, identifier > 0)
-    position   <- rank(-tabulate(values), ties = 'first')
-    position   <- factor(values, seq(max(values)), position)
-    position   <- as.integer(levels(position))[position]
-    identifier <- replace(identifier, identifier > 0, position)
-    return(identifier)
+  values     <- subset(identifier, identifier > 0)
+  position   <- rank(-tabulate(values), ties = 'first')
+  position   <- position[match(values, seq(max(values)))]
+  identifier <- replace(identifier, identifier > 0, position)
+  return(identifier)
 })
 
 # Computes identifiers
@@ -224,70 +247,72 @@ identifier_foo <- cmpfun(function(delineation, livable, params) {
   return(identifier)
 })
 
-# Computations ------------------------------------------------------------
+# Urban areas -------------------------------------------------------------
 
 tic('Runtime')
 
 # Urban areas data
-cat('\n- Loading data')
+cat('\n- Computing urban areas: data...')
 kernel  <- kernel_foo(params$bandwidth)
 livable <- read_foo(params$unlivable)
 livable <- replace(livable, px.na(livable), 1) == 0
-livable <- pad(livable, 2 * ceiling(params$bandwidth / 2), "xy")
-rescale <- replace(convolve(livable, kernel), !livable, 0)
+livable <- pad(livable, nPix = params$bandwidth, 'xy')
 density <- read_foo(params$density, livable)
 
 # Urban areas computations
-cat('\n- Computing urban areas:')
-params      <- optimise_foo(livable, params)                                ; cat(' bootstraps...')
-bootstraps  <- bootstraps_foo(density, livable, rescale, kernel, params)    ; cat(' thresholds...')
-threshold   <- threshold_foo(bootstraps, livable, params)                   ; cat(' delineations...')
-delineation <- delineation_foo(density, livable, rescale, threshold, kernel); cat(' identifiers')
-urban       <- identifier_foo(delineation, livable, params)
+params      <- optimise_foo(livable, params)                       ; cat(' bootstraps...')
+bootstraps  <- bootstraps_foo(density, livable, kernel, params)    ; cat(' thresholds...')
+threshold   <- threshold_foo(bootstraps, livable, params)          ; cat(' delineations...')
+delineation <- delineation_foo(density, livable, threshold, kernel); cat(' identifiers...')
+urban       <- identifier_foo(delineation, livable, params)        ; cat(' saving')
 
 # Writes urban areas
 write_foo(threshold, 'ut', params)
 write_foo(urban, 'ur', params, navalue = 0)
-
-if(params$usedisk) {
-  unlink(bootstraps, recursive = T) 
-}
+if(params$usedisk) unlink(bootstraps, recursive = T)
 rm(bootstraps, threshold, delineation)
 
-# Urban core(s) computations
-if(!exists("cores")) {
-  cores <- urban
-}
+# Urban cores -------------------------------------------------------------
+
+# Reference delineation
+previous <- urban
 
 for(iter in seq(params$niter)) {
-  
+  # First iteration is not indexed
+  corelab <- ifelse(iter == 1, '', as.character(iter))
+ 
   # Urban cores data
-  cat(sprintf('\n- Computing urban cores %i:', iter))
-  livable <- imsub(cores > 0)
-  rescale <- convolve(livable, kernel)
-  rescale <- replace(rescale, !livable, 0)
-  density <- replace(density, !livable, 0)
+  cat('\n- Computing urban cores: data...')
+  livable1 <- previous > 0
+  livable2 <- convolve(livable1, kernel) > 0 
+  livable2 <- replace(livable2, !livable | livable1, F)
+  livable  <- livable1 | livable2
+  density1 <- replace(density, !livable1, 0)
+  density2 <- replace(density, !livable2, 0)
   
   # Cores computations
-  params      <- optimise_foo(livable, params)                                ; cat(' bootstraps...')
-  bootstraps  <- bootstraps_foo(density, livable, rescale, kernel, params)    ; cat(' thresholds...')
-  threshold   <- threshold_foo(bootstraps, livable, params)                   ; cat(' delineations...')
-  delineation <- delineation_foo(density, livable, rescale, threshold, kernel); cat(' identifiers')
-  cores       <- identifier_foo(delineation, livable, params)
+  params      <- optimise_foo(livable, params)                                                ; cat(' bootstraps...')
+  bootstraps  <- bootstraps_cores_foo(density1, livable1, density2, livable2, kernel, params) ; cat(' thresholds...')
+  threshold   <- threshold_foo(bootstraps, livable, params)                                   ; cat(' delineations...')
+  delineation <- delineation_foo(density, livable, threshold, kernel)                         ; cat(' identifiers...')
+  delineation <- replace(delineation, previous == 0, 0)
+  cores       <- identifier_foo(delineation, livable, params)                                 ; cat(' saving')
   
   # Writes cores
-  write_foo(threshold, sprintf('ct%i', iter), params)
-  write_foo(cores, sprintf('co%i', iter), params, navalue = 0)
-  if(params$usedisk) {
-    unlink(bootstraps, recursive = T)  
-  }
-  rm(bootstraps, threshold, delineation)
+  write_foo(threshold, sprintf('ct%s', corelab), params)
+  write_foo(cores, sprintf('co%s', corelab), params, navalue = 0)
+  if(params$usedisk) unlink(bootstraps, recursive = T)  
+  rm(bootstraps, threshold, delineation, density1, density2, livable1, livable2)
   
   # Urban areas with cores
-  cat(sprintf('\n- Computing urban areas with cores %i\n\n', iter))
-  urbancores <- replace(urban, !(urban %in% unique(subset(urban, cores > 0))), 0)
-  write_foo(urbancores, sprintf('cc%i', iter), params, navalue = 0)
-
+  cat('\n- Computing urban areas with cores\n\n')
+  urbancores <- urban %in% unique(subset(urban, cores > 0))
+  urbancores <- replace(urban, !urbancores, 0)
+  write_foo(urbancores, sprintf('cc%s', corelab), params, navalue = 0)
+  rm(corelab)
+  
+  # Updates reference
+  previous <- cores
 }
 
 toc()
